+++ kernel/os/src/os_sem.c
@@ -35,7 +35,7 @@ os_sem_init(struct os_sem *sem, uint16_t tokens)
 {
     os_error_t ret;
 
-    os_trace_api_u32x2(OS_TRACE_ID_SEM_INIT, (uint32_t)sem, (uint32_t)tokens);
+    os_trace_api_u32x2(OS_TRACE_ID_SEM_INIT, (uintptr_t)sem, (uintptr_t)tokens);
 
     if (!sem) {
         ret = OS_INVALID_PARM;
@@ -48,7 +48,7 @@ os_sem_init(struct os_sem *sem, uint16_t tokens)
     ret = OS_OK;
 
 done:
-    os_trace_api_ret_u32(OS_TRACE_ID_SEM_INIT, (uint32_t)ret);
+    os_trace_api_ret_u32(OS_TRACE_ID_SEM_INIT, (uintptr_t)ret);
     return ret;
 }
 
@@ -61,7 +61,7 @@ os_sem_release(struct os_sem *sem)
     struct os_task *rdy;
     os_error_t ret;
 
-    os_trace_api_u32(OS_TRACE_ID_SEM_RELEASE, (uint32_t)sem);
+    os_trace_api_u32(OS_TRACE_ID_SEM_RELEASE, (uintptr_t)sem);
 
     /* OS must be started to release semaphores */
     if (!g_os_started) {
@@ -115,7 +115,7 @@ os_sem_release(struct os_sem *sem)
     ret = OS_OK;
 
 done:
-    os_trace_api_ret_u32(OS_TRACE_ID_SEM_RELEASE, (uint32_t)ret);
+    os_trace_api_ret_u32(OS_TRACE_ID_SEM_RELEASE, (uintptr_t)ret);
     return ret;
 }
 
@@ -129,7 +129,7 @@ os_sem_pend(struct os_sem *sem, os_time_t timeout)
     struct os_task *last;
     os_error_t ret;
 
-    os_trace_api_u32x2(OS_TRACE_ID_SEM_PEND, (uint32_t)sem, (uint32_t)timeout);
+    os_trace_api_u32x2(OS_TRACE_ID_SEM_PEND, (uintptr_t)sem, (uintptr_t)timeout);
 
     /* Check if OS is started */
     if (!g_os_started) {
@@ -203,6 +203,6 @@ os_sem_pend(struct os_sem *sem, os_time_t timeout)
     }
 
 done:
-    os_trace_api_ret_u32(OS_TRACE_ID_SEM_PEND, (uint32_t)ret);
+    os_trace_api_ret_u32(OS_TRACE_ID_SEM_PEND, (uintptr_t)ret);
     return ret;
 }
